LiveText updated for new timeObj class.
The new time obj does NOT auto start. I had to put the call to start into this so it'll work again. I put it the addAColor() method. It'll never be called before start(), and typically never used during runtime.

Also moved hookup() in there as well, for all the same reasons.